Improvements and bug fixes *Using aiohttp library to make async requests to EvilInsult API *Refactored send_warning() *Refactored send_command_usage() *Added exception WrongChannelError to handle if the user tries to use the bot's commands in the wrong channel *Sending warnings about unhandled command errors *Simplified listener on_command_completion() *Optimized command 'clear' using TextChannel.purge() method *Minor changes and optimizations

+++ main.py
@@ -2,11 +2,9 @@
 import discord
 import os
 import typing
-import asyncio
+import aiohttp
 from discord.ext import commands
-from discord.ext.commands import errors
 from pretty_help import PrettyHelp
-from urllib import request
 import music_commands
 
 description = 'A moderator bot for your server.'
@@ -25,44 +23,36 @@ ch_commands = None
 ch_bot_test = None
 
 
-async def send_warning(message, warning_type, title, description):
-  formatted_description = ':no_entry_sign:   **{0}**\n\n{1}'.format(title, description)
-  warning = discord.Embed(title=warning_type, description=formatted_description, color = 0xff0000)
-
-  await message.reply(embed=warning, delete_after=5)
-  await asyncio.sleep(5)
-  await message.delete()
+class WrongChannelError(commands.CommandError):
+  'Raised when the user wants to run a command in the channel of music bot.'
+  pass
 
 
-async def send_command_usage(ctx):
-  names = [ctx.command.name]
-  names += ctx.command.aliases
-  names = str(names).replace('\'', '').replace(', ', '|')
-  signature = ctx.command.signature
-  usage = f'```{bot.command_prefix}{names} {signature}```'
-  
-  embed = discord.Embed(title=ctx.command.name)
+async def send_warning(reply_to, title, sub_title, description, delete_after=None):
+  'Sends a custom warning by replying to a message.'
+  formatted_description = (
+    f':no_entry_sign:\t**{sub_title}**\n\n{description}'
+  )
+  warning = discord.Embed(title=title, description=formatted_description, color=0xff0000)
+  return await reply_to.reply(embed=warning, delete_after=delete_after)
+
+
+async def send_command_usage(reply_to, command, delete_after=None):
+  'Sends a command\'s usage by replying to a message.'
+  names = '|'.join([command.name] + command.aliases)
+  signature = command.signature
+  usage = f'```{bot.command_prefix}[{names}] {signature}```'
+
+  embed = discord.Embed(title=command.name)
   embed.add_field(name='Usage', value=usage, inline=True)
-  await ctx.message.delete()
-  await ctx.send(embed=embed, delete_after=15)
+  return await reply_to.reply(embed=embed, delete_after=delete_after)
 
 
-def extract_args(ctx):
-  args = ''
-  for arg in ctx.args:
-    if not arg is ctx.args[0]:
-      if not arg:
-        continue
-        
-      if isinstance(arg, discord.TextChannel):
-        args += f'#{arg} '
-      elif isinstance(arg, discord.User):
-        args += f'@{arg} '
-      elif isinstance(arg, str):
-        args += f'\'{arg}\' '
-      else:
-        args += f'{arg} '
-  return args
+async def get_random_insult():
+  'Gets a random insult asynchronously using the EvilInsult API'
+  async with aiohttp.ClientSession() as session:
+    async with session.get('https://evilinsult.com/generate_insult.php') as response:
+      return await response.text()
 
 
 @bot.event
@@ -79,55 +69,70 @@ async def on_ready():
   ch_commands = bot.get_channel(827374609234001970)
   ch_bot_test = bot.get_channel(827374679928471592)
 
-  print('Logged in as', bot.user)
+  print('Logged in as', bot.user, flush=True)
   await ch_log.send(f'{bot.user.mention} is online! :vulcan:')
-
-
-@bot.event
-async def on_command_error(ctx, exception):
-  if isinstance(exception, errors.CommandNotFound):
-    title = 'Unknown command!'
-    description = 'The given command could not be found.'
-    await send_warning(ctx.message, 'Notice!', title, description)
-  elif isinstance(exception, discord.ext.commands.BadArgument):
-    await send_command_usage(ctx)
-  else:
-    print(exception)
 
 
 @bot.listen()
 async def on_message(message):
-  if not message.author.bot:
-    music_command = message.content.startswith(tuple(music_commands.commands))
-    if not message.channel is ch_music:
-      if music_command:
-        title = 'Wrong channel!'
-        description = 'You can only use {0.mention}\'s command in {1.mention}.'.format(music_bot, ch_music)
+  if not message.content.startswith(music_commands.command_prefix) or message.author.bot:
+    return
 
-        await send_warning(message, 'Warning!', title, description)
-    else:
-      if not music_command:
-        title = 'Prohibited!'
-        description = f'You can only send {music_bot.mention}\'s commands in this channel.'
-        await send_warning(message, 'Warning!', title, description)  
+  is_music_command = message.content.startswith(music_commands.commands)
+  if is_music_command and not message.channel is ch_music:
+    title = 'Warning!'
+    sub_title = 'Wrong channel!'
+    description = f'You can only use {music_bot.mention}\'s command in {ch_music.mention}.'
+    await send_warning(reply_to=message, title=title, sub_title=sub_title, description=description, delete_after=10)
+    await message.delete(delay=10)
+
+
+@bot.check
+async def check_channel(ctx):
+  if ctx.channel is ch_music:
+    raise WrongChannelError
+
+  return True
+
+
+@bot.event
+async def on_command_error(ctx, exception):
+  if isinstance(exception, commands.CommandNotFound):
+    title = 'Unknown command!'
+    description = 'The given command could not be found.'
+    await send_warning(reply_to=ctx.message, title='Notice!', sub_title=title, description=description, delete_after=10)
+
+  elif isinstance(exception, commands.BadArgument):
+    title = 'Error!'
+    sub_title = 'Bad argument!'
+    description = 'Command got a bad argument.'
+    await send_warning(reply_to=ctx.message, title=title, sub_title=sub_title, description=description, delete_after=10)
+    await send_command_usage(reply_to=ctx.message, command=ctx.command, delete_after=10)
+
+  elif isinstance(exception, WrongChannelError):
+    title = 'Warning!'
+    sub_title = 'Prohibited!'
+    description = f'You can only use {music_bot.mention}\'s commands in this channel.'
+    await send_warning(reply_to=ctx.message, title=title, sub_title=sub_title, description=description, delete_after=10)
+
+  else:
+    title = 'Error!'
+    sub_title = 'Unknown error!'
+    description = f'An unknown command error occured!\nException: {exception}'
+    await send_warning(reply_to=ctx.message, title=title, sub_title=sub_title, description=description, delete_after=10)
+  
+  await ctx.message.delete(delay=10)
 
 
 @bot.listen()
 async def on_command_completion(ctx):
   user = ctx.author
-  command = ctx.command
   channel = ctx.channel
-  args = extract_args(ctx).strip()
-  log_msg = f'{user.mention} used in {channel.mention}```${command} {args}```'.strip()
+  cmd_msg = ctx.message
+  log_msg = f'{user.mention} used a command in {channel.mention} \n```{cmd_msg.clean_content}```'
+
   embed = discord.Embed(description=log_msg)
-
   await ch_log.send(embed=embed, allowed_mentions=discord.AllowedMentions.none())
-
-
-@bot.check
-async def pre_check(ctx):
-  if not ctx.channel is ch_music:
-    return True
 
 
 @bot.command(name='ping')
@@ -139,11 +144,12 @@ async def _ping(ctx):
 @bot.command(name='nick', aliases=['nickname', 'name'])
 async def _nick(ctx, name : str = None):
   'Changes your name to whatever you want. You cannot have spaces in your name!'
-  if not name == None: 
+  if name: 
     await ctx.author.edit(nick=name)
   else:
-    await send_command_usage(ctx)
-    
+    await send_command_usage(reply_to=ctx.message, command=ctx.command, delete_after=10)
+    await ctx.message.delete(delay=10)
+
 
 @bot.command(name='insult', aliases=['bother', 'curse'])
 async def _insult(ctx, target_user : discord.User = None):
@@ -152,39 +158,43 @@ async def _insult(ctx, target_user : discord.User = None):
   if not target_user:
     target_user = ctx.author
 
-  random_insult = ''
-  # Denote the long time the request takes by 'typing' to the channel
+  # Denote the time consuming request by 'typing' to the channel
   async with ctx.typing():
-    response = request.urlopen('https://evilinsult.com/generate_insult.php')
-    random_insult = response.read().decode('UTF-8')
+    random_insult = await get_random_insult()
 
-  await ctx.send("{0.mention} {1}".format(target_user, random_insult))
+  msg = f'{target_user.mention} {random_insult}'
+  await ctx.message.reply(msg)
 
 
 @bot.command(name='pin')
 async def _pin(ctx):
-  'Pins the last message of the channel that wasn\'t sent by the bot itself.'
+  'Pins the last message of the channel that\'s not a command nor sent by the bot.'
   history=ctx.channel.history(limit=10)
-  predicate = lambda message: not message.content.startswith(bot.command_prefix) and not message.author == bot.user
+  predicate = lambda msg: not msg.content.startswith(ctx.prefix) and not msg.author.bot
   last_user_made_message = await history.find(predicate)
 
   if last_user_made_message:
     await last_user_made_message.pin()
   else:
-    await send_warning(ctx.message, 'Notice!', '', 'Couldn\'t find any pinnable messages in the last 10 entries.')
+    title = 'Notice!'
+    description = 'Couldn\'t find any pinnable messages in the last 10 entries.'
+    await send_warning(reply_to=ctx.message, title=title, sub_title='', description=description, delete_after=10)
+    await ctx.message.delete(delay=10)
 
 
-@bot.command(name='clear', aliases=['erase', 'clean'])
+@bot.command(name='clear', aliases=['erase', 'clean', 'purge'])
 async def _clear(ctx, channel : typing.Optional[discord.TextChannel], count : int = 1):
   'Deletes past messages based on given number.'
-  
-  await ctx.message.delete()
   if not channel:
     channel = ctx.channel
-    
-  history=channel.history(limit=count, oldest_first=False)
-  async for message in history:
-      await message.delete()
+
+  async with ctx.typing():
+    if channel is ctx.channel:
+      await ctx.message.delete()
+    deleted_messages = await channel.purge(limit=count)
+  
+  msg = f'Deleted {len(deleted_messages)} entries!'
+  await ctx.send(msg, delete_after=10)
 
 
 if __name__ == '__main__':
